+++ packages/@livestore/common/src/leader-thread/apply-mutation.ts
@@ -1,4 +1,4 @@
-import { env, memoizeByRef, shouldNeverHappen } from '@livestore/utils'
+import { memoizeByRef, shouldNeverHappen } from '@livestore/utils'
 import type { Scope } from '@livestore/utils/effect'
 import { Effect, Option, Schema } from '@livestore/utils/effect'
 
@@ -46,7 +46,7 @@ export const makeApplyMutation: Effect.Effect<ApplyMutation, never, Scope.Scope 
 
         // console.group('[@livestore/common:leader-thread:applyMutation]', { mutationName })
 
-        const session = env('VITE_LIVESTORE_EXPERIMENTAL_SYNC_NEXT') ? db.session() : undefined
+        const session = db.session()
 
         for (const { statementSql, bindValues } of execArgsArr) {
           // console.debug(mutationName, statementSql, bindValues)
@@ -54,26 +54,24 @@ export const makeApplyMutation: Effect.Effect<ApplyMutation, never, Scope.Scope 
           yield* execSqlPrepared(db, statementSql, bindValues)
         }
 
-        if (session !== undefined) {
-          const changeset = session.changeset()
-          session.finish()
-          // NOTE for no-op mutations (e.g. if the state didn't change) the changeset will be empty
-          // TODO possibly write a null value instead of omitting the row
-          if (changeset.length > 0) {
-            // TODO use prepared statements
-            yield* execSql(
-              db,
-              ...insertRow({
-                tableName: SESSION_CHANGESET_META_TABLE,
-                columns: sessionChangesetMetaTable.sqliteDef.columns,
-                values: {
-                  idGlobal: mutationEventEncoded.id.global,
-                  idLocal: mutationEventEncoded.id.local,
-                  changeset,
-                },
-              }),
-            )
-          }
+        const changeset = session.changeset()
+        session.finish()
+        // NOTE for no-op mutations (e.g. if the state didn't change) the changeset will be empty
+        // TODO possibly write a null value instead of omitting the row
+        if (changeset.length > 0) {
+          // TODO use prepared statements
+          yield* execSql(
+            db,
+            ...insertRow({
+              tableName: SESSION_CHANGESET_META_TABLE,
+              columns: sessionChangesetMetaTable.sqliteDef.columns,
+              values: {
+                idGlobal: mutationEventEncoded.id.global,
+                idLocal: mutationEventEncoded.id.local,
+                changeset,
+              },
+            }),
+          )
         }
 
         // console.groupEnd()