+++ packages/@livestore/node/src/leader-thread.ts
@@ -10,14 +10,13 @@ if (process.execArgv.includes('--inspect')) {
 
 import { NodeFileSystem, NodeWorkerRunner } from '@effect/platform-node'
 import type { NetworkStatus } from '@livestore/common'
-import { Devtools, ROOT_ID, sql, UnexpectedError } from '@livestore/common'
+import { Devtools, MutationEventEncodedWithDeferred, ROOT_ID, sql, UnexpectedError } from '@livestore/common'
 import type { DevtoolsContext } from '@livestore/common/leader-thread'
 import {
   configureConnection,
   LeaderThreadCtx,
   makeDevtoolsContext,
   makeLeaderThreadLayer,
-  MutationEventEncodedWithDeferred,
 } from '@livestore/common/leader-thread'
 import type { LiveStoreSchema } from '@livestore/common/schema'
 import { makeNodeDevtoolsChannel } from '@livestore/devtools-node-common/web-channel'